Add files via upload
change the write class so that it can plot the loss and accuracy correctly.  Now, the abscissa represents time instead the . The mistake in loading the validate data is fixed.  The index of class presents the sequence of the files in the train directory, instead of the sequence in the wnids.txt :(